+++ mypyc/emitclass.py
@@ -12,7 +12,7 @@ from mypyc.emitwrapper import (
     generate_bool_wrapper, generate_get_wrapper,
 )
 from mypyc.rtypes import RType, RTuple, object_rprimitive
-from mypyc.func_ir import FuncIR, FuncDecl, FUNC_STATICMETHOD, FUNC_CLASSMETHOD
+from mypyc.ir.func_ir import FuncIR, FuncDecl, FUNC_STATICMETHOD, FUNC_CLASSMETHOD
 from mypyc.class_ir import ClassIR, VTableMethod, VTableEntries
 from mypyc.sametype import is_same_type
 from mypyc.namegen import NameGenerator
