+++ mypyc/emitwrapper.py
@@ -9,7 +9,7 @@ from mypyc.emit import Emitter
 from mypyc.rtypes import (
     RType, is_object_rprimitive, is_int_rprimitive, is_bool_rprimitive, object_rprimitive
 )
-from mypyc.func_ir import FuncIR, RuntimeArg, FUNC_STATICMETHOD
+from mypyc.ir.func_ir import FuncIR, RuntimeArg, FUNC_STATICMETHOD
 from mypyc.class_ir import ClassIR
 from mypyc.namegen import NameGenerator
