+++ mypyc/exceptions.py
@@ -15,7 +15,7 @@ from mypyc.ir.ops import (
     BasicBlock, LoadErrorValue, Return, Branch, RegisterOp, ERR_NEVER, ERR_MAGIC,
     ERR_FALSE, NO_TRACEBACK_LINE_NO,
 )
-from mypyc.func_ir import FuncIR
+from mypyc.ir.func_ir import FuncIR
 
 
 def insert_exception_handling(ir: FuncIR) -> None:
