+++ mypyc/irbuild/ll_builder.py
@@ -26,7 +26,7 @@ from mypyc.rtypes import (
     RType, RUnion, RInstance, optional_value_type, int_rprimitive, float_rprimitive,
     bool_rprimitive, list_rprimitive, str_rprimitive, is_none_rprimitive, object_rprimitive
 )
-from mypyc.func_ir import FuncDecl, FuncSignature
+from mypyc.ir.func_ir import FuncDecl, FuncSignature
 from mypyc.class_ir import ClassIR, all_concrete_classes
 from mypyc.common import (
     FAST_ISINSTANCE_MAX_SUBCLASSES, MAX_LITERAL_SHORT_INT,
