+++ mypyc/module_ir.py
@@ -5,7 +5,7 @@ from typing import List, Tuple, Dict
 from mypyc.common import JsonDict
 from mypyc.ir.ops import DeserMaps
 from mypyc.rtypes import RType, deserialize_type
-from mypyc.func_ir import FuncIR, FuncDecl, format_func
+from mypyc.ir.func_ir import FuncIR, FuncDecl, format_func
 from mypyc.class_ir import ClassIR
