+++ mypyc/refcount.py
@@ -30,7 +30,7 @@ from mypyc.ir.ops import (
     BasicBlock, Assign, RegisterOp, DecRef, IncRef, Branch, Goto, Environment,
     Op, ControlOp, Value, Register
 )
-from mypyc.func_ir import FuncIR
+from mypyc.ir.func_ir import FuncIR
 
 
 DecIncs = Tuple[Tuple[Tuple[Value, bool], ...], Tuple[Value, ...]]
