+++ mypyc/uninit.py
@@ -11,7 +11,7 @@ from mypyc.analysis import (
 from mypyc.ir.ops import (
     BasicBlock, Branch, Value, RaiseStandardError, Unreachable, Environment, Register
 )
-from mypyc.func_ir import FuncIR
+from mypyc.ir.func_ir import FuncIR
 
 
 def insert_uninit_checks(ir: FuncIR) -> None:
